# contracts/openapi/host.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Permguard Host API",
    "version": "1",
    "description": "The Host listener (WP-2.5): `/host/v1` over TLS on `admin.addr`, and `permguard.host.v1` on the same port, one facade behind both. Every route authorizes with the Host's grants; every mutation carries a request id, may state the revision it expects, and answers a receipt. Identity, membership, evidence and the key and configuration mutations arrive with their packages; this document describes what this build serves."
  },
  "paths": {
    "/host/v1/identity": {
      "get": {
        "operationId": "getHostIdentity",
        "summary": "The Host identity",
        "description": "Under `identity.read` on the Host. The public identity: the signed document, the succession records, the epoch-1 public key and the protocol versions.",
        "responses": {
          "200": {
            "description": "The Host identity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostIdentity"
                }
              }
            }
          },
          "401": {
            "description": "No usable credential: nobody, or a credential no rule maps. Carries `WWW-Authenticate: Mutual-TLS realm=\"permguard\"`.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "common.json#/components/schemas/WireDenial"
                }
              }
            }
          },
          "403": {
            "description": "A principal without the grant the route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "common.json#/components/schemas/WireDenial"
                }
              }
            }
          },
          "503": {
            "description": "No Host identity is open (`identity_unavailable`), or the principal is over its bound (`principal_bound_exceeded`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostWireError"
                }
              }
            }
          }
        },
        "x-grpc-rpc": "permguard.host.v1.IdentityService/GetIdentity"
      }
    },
    "/host/v1/identity/rotate": {
      "post": {
        "operationId": "rotateHostIdentity",
        "summary": "Rotate the Host identity",
        "description": "Under `identity.admin` on the Host. A security mutation: the next epoch's key is generated, the current key signs the succession, the next signs the document. A retry with the same request id answers the stored receipt.",
        "requestBody": {
          "required": true,
          "description": "The envelope's request id and the epoch the caller read.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RotateIdentityBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rotated; a retry with the same request id answers the stored receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityRotated"
                }
              }
            }
          },
          "400": {
            "description": "The body is not the request's shape (`invalid_argument`); the request id is absent or out of bounds (`request_id_required`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostWireError"
                }
              }
            }
          },
          "401": {
            "description": "No usable credential: nobody, or a credential no rule maps. Carries `WWW-Authenticate: Mutual-TLS realm=\"permguard\"`.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "common.json#/components/schemas/WireDenial"
                }
              }
            }
          },
          "403": {
            "description": "A principal without the grant the route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "common.json#/components/schemas/WireDenial"
                }
              }
            }
          },
          "409": {
            "description": "The epoch is not the one the caller read (`revision_mismatch`, the current epoch beside it), or the request id is in flight or was used for another request (`request_id_reused`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostWireError"
                }
              }
            }
          },
          "500": {
            "description": "The mutation was applied, or may have been, and could not be recorded (`mutation_unrecorded`): read the current state before retrying, or retry with the same `request_id`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostWireError"
                }
              }
            }
          },
          "503": {
            "description": "No Host identity is open or it could not rotate (`identity_unavailable`), the mutation journal or the audit trail cannot record (`replay_unavailable`, `audit_unavailable`, nothing applied), or the principal is over its bound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostWireError"
                }
              }
            }
          }
        },
        "x-grpc-rpc": "permguard.host.v1.IdentityService/RotateIdentity"
      }
    },
    "/host/v1/identity/reset/plan": {
      "post": {
        "operationId": "planIdentityReset",
        "summary": "Plan an identity reset",
        "description": "Under `identity.admin` on the Host (WP-4.1). The first step of a reset: the plan names every live membership and what the reset does to it, and is held by the server, consumed once by the run.",
        "requestBody": {
          "required": true,
          "description": "The envelope's request id, the mode and the reason.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanIdentityResetBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The plan, to present to the run step within its lifetime.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResetPlan"
                }
              }
            }
          },
          "400": {
            "description": "The body is not the request's shape (`invalid_argument`); the request id is absent or out of bounds (`request_id_required`); the mode or reason is malformed (`invalid_argument`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostWireError"
                }
              }
            }
          },
          "401": {
            "description": "No usable credential: nobody, or a credential no rule maps. Carries `WWW-Authenticate: Mutual-TLS realm=\"permguard\"`.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "common.json#/components/schemas/WireDenial"
                }
              }
            }
          },
          "403": {
            "description": "A principal without the grant the route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "common.json#/components/schemas/WireDenial"
                }
              }
            }
          },
          "409": {
            "description": "The request id was reused (`request_id_reused`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostWireError"
                }
              }
            }
          },
          "500": {
            "description": "The plan may have been held and could not be recorded (`mutation_unrecorded`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostWireError"
                }
              }
            }
          },
          "503": {
            "description": "No identity is open, it was reset already, or this process cannot provision another (`identity_unavailable`); the journal or audit trail cannot record (`audit_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostWireError"
                }
              }
            }
          }
        },
        "x-grpc-rpc": "permguard.host.v1.IdentityService/PlanIdentityReset"
      }
    },
    "/host/v1/identity/reset/run": {
      "post": {
        "operationId": "runIdentityReset",
        "summary": "Reset the identity",
        "description": "Under `identity.admin` on the Host (WP-4.1). Each coordinator asked to revoke the memberships this Host is a member of; the memberships it coordinates ended here; the identity retired, its public evidence kept under `host/identity/retired/`, its private keys destroyed and a new identity provisioned. The process signs nothing more and serves the new identity after a restart.",
        "requestBody": {
          "required": true,
          "description": "The envelope's request id and the plan, as the plan step answered it.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunIdentityResetBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The receipt, the memberships ended and orphaned, and the new identity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityReset"
                }
              }
            }
          },
          "400": {
            "description": "The body is not the request's shape (`invalid_argument`); the request id is absent or out of bounds (`request_id_required`); the plan digest is not the one answered (`plan_digest_mismatch`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostWireError"
                }
              }
            }
          },
          "401": {
            "description": "No usable credential: nobody, or a credential no rule maps. Carries `WWW-Authenticate: Mutual-TLS realm=\"permguard\"`.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "common.json#/components/schemas/WireDenial"
                }
              }
            }
          },
          "403": {
            "description": "A principal without the grant the route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "common.json#/components/schemas/WireDenial"
                }
              }
            }
          },
          "404": {
            "description": "No reset plan of that id is held (`plan_unknown`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostWireError"
                }
              }
            }
          },
          "409": {
            "description": "A normal reset some coordinator did not acknowledge, the identity unchanged (`identity_reset_incomplete`); the identity or a membership moved since the plan (`revision_mismatch`); the plan expired or ran (`plan_expired`); or the request id was reused (`request_id_reused`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostWireError"
                }
              }
            }
          },
          "500": {
            "description": "The reset began and did not complete (`mutation_unrecorded`): run it again, online or with the offline CLI.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostWireError"
                }
              }
            }
          },
          "503": {
            "description": "No identity is open, it was reset already, or this process cannot provision another (`identity_unavailable`); the journal or audit trail cannot record (`audit_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostWireError"
                }
              }
            }
          }
        },
        "x-grpc-rpc": "permguard.host.v1.IdentityService/RunIdentityReset"
      }
    },
    "/host/v1/ring-bindings": {
      "get": {
        "operationId": "listRingBindings",
        "summary": "The signed ring bindings",
        "description": "Under `identity.read` on the Host: the identity-signed binding of every ring's current epoch (WP-3.1). `host.identity` has none: its chain authenticates it.",
        "responses": {
          "200": {
            "description": "Every binding held, one per ring that has keys of its own.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RingBindings"
                }
              }
            }
          },
          "401": {
            "description": "No usable credential: nobody, or a credential no rule maps. Carries `WWW-Authenticate: Mutual-TLS realm=\"permguard\"`.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "common.json#/components/schemas/WireDenial"
                }
              }
            }
          },
          "403": {
            "description": "A principal without the grant the route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "common.json#/components/schemas/WireDenial"
                }
              }
            }
          },
          "503": {
            "description": "A ring could not be read (`ring_unreadable`), or the principal is over its bound (`principal_bound_exceeded`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostWireError"
                }
              }
            }
          }
        },
        "x-grpc-rpc": "permguard.host.v1.IdentityService/ListRingBindings"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HostIdentity": {
        "type": "object",
        "properties": {
          "document": {
            "type": "string",
            "description": "The signed identity document of the current epoch: COSE_Sign1 `permguard.host.identity.v1`, base64url without padding.",
            "x-grpc-field": "document"
          },
          "successions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Every succession record, oldest first: COSE_Sign1 `permguard.host.succession.v1` each, base64url without padding.",
            "x-grpc-field": "successions"
          },
          "first_public_key": {
            "type": "string",
            "description": "The epoch-1 public key, the suite's raw bytes, base64url without padding: what the first fingerprint pins, from which the chain verifies.",
            "x-grpc-field": "first_public_key"
          },
          "protocol_versions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The protocol versions this Host speaks.",
            "x-grpc-field": "protocol_versions"
          }
        },
        "required": [
          "document",
          "successions",
          "first_public_key",
          "protocol_versions"
        ],
        "additionalProperties": false,
        "description": "The public Host identity (WP-2.2).",
        "x-grpc-message": "permguard.host.v1.GetIdentityResponse"
      },
...
    }
  },
...
}